Fix startup_policy "None Type" errors
From libvirt 2.0.0 on, the disk source element is unable
to check under error phenomenons. Fix by check disk source
according to libvirt version.

+++ libvirt/tests/src/virtual_disks/startup_policy.py
@@ -1,67 +1,11 @@
 import os
 import logging
 import shutil
-
 from autotest.client import utils
-from autotest.client.shared import error
-
 from virttest import virsh
 from virttest.libvirt_xml import vm_xml
 from virttest.utils_test import libvirt
-
-
-def create_disk_xml(xml_file, device_type, source_file, target_dev, policy):
-    """
-    Create a disk xml file for attaching to a domain.
-
-    :prams xml_file: path/file to save the disk XML
-    :source_file: disk's source file
-    :device_type: CD-ROM or floppy
-    """
-    if device_type == "cdrom":
-        target_bus = "ide"
-        image_size = "100M"
-    elif device_type == "floppy":
-        target_bus = "fdc"
-        image_size = "1.44M"
-    else:
-        error.TestNAError("Unsupport device type in this test: " + device_type)
-    utils.run("qemu-img create %s %s" % (source_file, image_size))
-    disk_class = vm_xml.VMXML.get_device_class('disk')
-    disk = disk_class(type_name='file')
-    disk.device = device_type
-    disk.driver = dict(name='qemu')
-    disk.source = disk.new_disk_source(attrs={'file': source_file, 'startupPolicy': policy})
-    disk.target = dict(bus=target_bus, dev=target_dev)
-    disk.xmltreefile.write()
-    shutil.copyfile(disk.xml, xml_file)
-
-
-def check_disk_source(vm_name, target_dev, expect_value):
-    """
-    Check the disk source: file and startupPolicy.
-
-    :params vm_name: Domain name
-    :params target_dev: Disk's target device
-    :params expect_value: Expect value of source file and source startupPolicy
-    """
-    logging.debug("Expect source file is '%s'", expect_value[0])
-    logging.debug("Expect source startupPolicy is '%s'", expect_value[1])
-    vmxml = vm_xml.VMXML.new_from_dumpxml(vm_name)
-    disks = vmxml.get_disk_all()
-    source_value = []
-    try:
-        disk_source = disks[target_dev].find('source')
-        source_value.append(disk_source.get('file'))
-        source_value.append(disk_source.get('startupPolicy'))
-    except KeyError:
-        raise error.TestError("No %s in domain %s" % (target_dev, vm_name))
-    logging.debug("Actual source file is '%s'", source_value[0])
-    logging.debug("Actual source startupPolicy is '%s'", source_value[1])
-    if source_value == expect_value:
-        logging.debug("Domain disk XML check pass")
-    else:
-        raise error.TestFail("Domain disk XML check fail")
+from provider import libvirt_version
 
 
 def run(test, params, env):
@@ -91,7 +35,76 @@ def run(test, params, env):
         start_error = False
         restore_error = False
     else:
-        error.TestNAError("Unsupport startupPolicy ''%s' in this test" % startup_policy)
+        test.cancel("Unsupport startupPolicy '%s' in this test" % startup_policy)
+
+    def create_disk_xml(xml_file, device_type, source_file, target_dev, policy):
+        """
+        Create a disk xml file for attaching to a domain.
+
+        :prams xml_file: path/file to save the disk XML
+        :source_file: disk's source file
+        :device_type: CD-ROM or floppy
+        """
+        if device_type == "cdrom":
+            target_bus = "ide"
+            image_size = "100M"
+        elif device_type == "floppy":
+            target_bus = "fdc"
+            image_size = "1.44M"
+        else:
+            test.cancel("Unsupport device type in this test: " + device_type)
+        utils.run("qemu-img create %s %s" % (source_file, image_size))
+        disk_class = vm_xml.VMXML.get_device_class('disk')
+        disk = disk_class(type_name='file')
+        disk.device = device_type
+        disk.driver = dict(name='qemu')
+        disk.source = disk.new_disk_source(attrs={'file': source_file, 'startupPolicy': policy})
+        disk.target = dict(bus=target_bus, dev=target_dev)
+        disk.xmltreefile.write()
+        shutil.copyfile(disk.xml, xml_file)
+
+    def check_disk_source(vm_name, target_dev, expect_value):
+        """
+        Check the disk source: file and startupPolicy.
+
+        :params vm_name: Domain name
+        :params target_dev: Disk's target device
+        :params expect_value: Expect value of source file and source startupPolicy
+        """
+        logging.debug("Expect source file is '%s'", expect_value[0])
+        logging.debug("Expect source startupPolicy is '%s'", expect_value[1])
+        vmxml = vm_xml.VMXML.new_from_dumpxml(vm_name)
+        disks = vmxml.get_disk_all()
+        source_value = []
+        try:
+            disk_source = disks[target_dev].find('source')
+            source_value.append(disk_source.get('file'))
+            source_value.append(disk_source.get('startupPolicy'))
+        except KeyError:
+            test.error("No %s in domain %s" % (target_dev, vm_name))
+        logging.debug("Actual source file is '%s'", source_value[0])
+        logging.debug("Actual source startupPolicy is '%s'", source_value[1])
+        if source_value == expect_value:
+            logging.debug("Domain disk XML check pass")
+        else:
+            test.fail("Domain disk XML check fail")
+
+    def rename_file(revert=False):
+        """
+        Rename a file or revert it.
+        """
+        try:
+            if not revert:
+                os.rename(media_file, media_file_new)
+                logging.debug("Rename %s to %s", media_file, media_file_new)
+            else:
+                os.rename(media_file_new, media_file)
+                logging.debug("Rename %s to %s", media_file_new, media_file)
+        except OSError, err:
+            test.fail("Rename image failed: %s" % str(err))
+
+    save_file = os.path.join(test.tmpdir, "vm.save")
+    expect_value = [None, startup_policy]
 
     # Back VM XML
     vmxml_backup = vm_xml.VMXML.new_from_inactive_dumpxml(vm_name)
@@ -107,29 +120,12 @@ def run(test, params, env):
     virsh_dargs = {'debug': True, 'ignore_status': True}
     if vm.is_alive():
         vm.destroy()
-    try:
-        virsh.attach_device(domainarg=vm_name, filearg=disk_xml_file,
-                            flagstr="--config", **virsh_dargs)
-    except:
+    ret = virsh.attach_device(domainarg=vm_name, filearg=disk_xml_file,
+                              flagstr="--config", **virsh_dargs)
+    if ret.exit_status:
         os.remove(media_file)
-        raise error.TestError("Attach %s fail", device_type)
+        test.error("Attach %s fail" % device_type)
 
-    def rename_file(revert=False):
-        """
-        Rename a file or revert it.
-        """
-        try:
-            if not revert:
-                os.rename(media_file, media_file_new)
-                logging.debug("Rename %s to %s", media_file, media_file_new)
-            else:
-                os.rename(media_file_new, media_file)
-                logging.debug("Rename %s to %s", media_file_new, media_file)
-        except OSError, err:
-            raise error.TestFail("Rename image failed: %s" % str(err))
-
-    save_file = os.path.join(test.tmpdir, "vm.save")
-    expect_value = [None, startup_policy]
     try:
         # Step 1. Start domain and destroy it normally
         vm.start()
@@ -139,8 +135,6 @@ def run(test, params, env):
         rename_file()
         result = virsh.start(vm_name, **virsh_dargs)
         libvirt.check_exit_status(result, expect_error=start_error)
-        if not start_error:
-            check_disk_source(vm_name, target_dev, expect_value)
 
         # Step 3. Move back the source file and start the domain(if needed)
         rename_file(revert=True)
@@ -153,8 +147,12 @@ def run(test, params, env):
         rename_file()
         result = virsh.restore(save_file, **virsh_dargs)
         libvirt.check_exit_status(result, expect_error=restore_error)
+
         if not restore_error:
-            check_disk_source(vm_name, target_dev, expect_value)
+            if not libvirt_version.version_compare(2, 0, 0):
+                check_disk_source(vm_name, target_dev, expect_value)
+            else:
+                logging.debug("Current libvirt version doesn't support disk source check")
 
         # Step 5. Move back the source file and restore the domain(if needed)
         rename_file(revert=True)
